Create Errors
This is just a rough draft of the error stack trace. I had to hard code throwing a random exception to print out the trace. Then it returns back to the mainMenu. I'm pretty sure this isn't the finalized code but it works nonetheless :) Enjoys.

I will work on it more later tonight but this is my progress!